Enhance top warp hint
